requirements.txt
pyTelegramBotAPI==4.27.0
psycopg2-binary>=2.9.10,<3.0
python-dotenv==1.0.1